requirements.txt
aiofiles==0.6.0
amqp==5.0.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
billiard==3.6.3.0
boto3==1.15.17
botocore==1.18.17
CacheControl==0.12.6
cachetools==4.1.1
celery==5.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-embed-video==1.3.3
django-extra-fields==2.0.5
django-google-api==1.0.1
django-notifs==2.6.3
django-resized==0.3.11
django-rest-passwordreset==1.1.0
django-simpleui==3.9.1
django-storages==1.10.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-yasg==1.17.1
ecdsa==0.14.1
enum-compat==0.0.3
eventlet==0.29.1
firebase-admin==4.4.0
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpcio==1.32.0
gunicorn==20.0.4
h11==0.9.0
httpcore==0.11.1
httplib2==0.18.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.0.2
MarkupSafe==1.1.1
monotonic==1.5
msgpack==1.0.0
multidict==5.0.0
mysqlclient==2.0.1
oauthlib==3.1.0
packaging==20.4
pika==1.1.0
Pillow==7.2.0
prompt-toolkit==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-jose==3.2.0
python-socketio==4.6.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
Sanic-Plugins-Framework==0.9.4.post1
six==1.15.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
swagger-spec-validator==2.7.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
websockets==8.1
whitenoise==5.2.0
wincertstore==0.2
yarl==1.6.2
zipp==3.3.1
